[김효정] sprint4 (#229)
* spring프로젝트 초기화

* refactor(mentor): sprint2 심화 요구 사항-관심사 분리를 통한 레이어 간 의존성 주입 추가. Basic*Service를 구현하여
Repository를 의존하게 설계하였습니다.

* 기본 요구 사항 완료

* 새로운 도메인 추가하기

* UserService 고도화

* AuthService 구현

* BinaryContentService, ReadStatusService 구현, ChannelService 일부 구현

* UserStatusService구현

* ChannelService 고도화

* MessageService 고도화

* MessageRepository 구현

* JCFRepository 구현

* 테스팅 및 순환의존성 수정 필요

* BasicChannelService와 BasicReadStatusService간 순환의존성 제거

* 백업, ReadStatus 수정필요(load의 초기화문제)

* 테스팅 필요(미완성), 스프린트4분기

* User API 테스트 1

* 사용자 관리, 권한 관리 완료

* channel에 속한 userList 서치로직구현

* refactor(mento): sprint3 코드리뷰 피드백 반영

* 채널 관리 완료

* 메시지 관리

* 메시지 수신 상태

* 바이너리 파일 상태

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicChannelService.java
@@ -3,6 +3,7 @@ package com.sprint.mission.discodeit.service.basic;
 import com.sprint.mission.discodeit.dto.channel.ChannelCreateDTO;
 import com.sprint.mission.discodeit.dto.channel.ChannelFindDTO;
 import com.sprint.mission.discodeit.dto.channel.ChannelUpdateDTO;
+import com.sprint.mission.discodeit.dto.channel.PrivateChannelCreateDTO;
 import com.sprint.mission.discodeit.dto.readstatus.ReadStatusCreateDTO;
 import com.sprint.mission.discodeit.dto.user.UserFindDTO;
 import com.sprint.mission.discodeit.entity.Channel;
@@ -37,7 +38,13 @@ public class BasicChannelService implements ChannelService {
     private final ReadStatusService readStatusService;
 
     @Override
-    public Channel createPrivateChannel(ChannelCreateDTO channelCreateDTO, ChannelType type) {
+    public Channel createPublicChannel(ChannelCreateDTO channelCreateDTO) {
+        Channel channel = new Channel(channelCreateDTO, ChannelType.PUBLIC);
+        return channelRepository.save(channel);
+    }
+
+    @Override
+    public Channel createPrivateChannel(PrivateChannelCreateDTO channelCreateDTO) {
         Channel channel = new Channel(channelCreateDTO, ChannelType.PRIVATE);
 
         createReadStatus(channel, channelCreateDTO);
@@ -45,20 +52,14 @@ public class BasicChannelService implements ChannelService {
         return channelRepository.save(channel);
     }
 
-    @Override
-    public Channel createPublicChannel(ChannelCreateDTO channelCreateDTO, ChannelType type) {
-        Channel channel = new Channel(channelCreateDTO, ChannelType.PUBLIC);
 
-        createReadStatus(channel, channelCreateDTO);
-        return channelRepository.save(channel);
-    }
 
     //ReadStatus서비스에서 ReadStatus를 만드는 함수
-    private void createReadStatus(Channel channel, ChannelCreateDTO channelCreateDTO) {
-        List<UUID> userIDList = channelCreateDTO.userList();
+    private void createReadStatus(Channel channel, PrivateChannelCreateDTO channelCreateDTO) {
+        List<UUID> userIDList = channelCreateDTO.getUserList();
         for(UUID uuid :userIDList){
             System.out.println("ReadStatus created");
-            readStatusService.create(new ReadStatusCreateDTO(channel.getId(), uuid ), isChannelExist(channel.getId()));
+            readStatusService.create(new ReadStatusCreateDTO(channel.getId(), uuid ));
         }
     }
 
@@ -73,12 +74,15 @@ public class BasicChannelService implements ChannelService {
     @Override
     public ChannelFindDTO findDTO(UUID uuid) {
         Channel channel = readChannel(uuid);
-        Instant time = readStatusRepository.findLatestTimeByChannelId(uuid);
 
-        //private일 때 userIdList 생성, Public시 null list
+
+        //Public 일 때 userIdList , time은 null
+        //이 아이디리스트는 readStatus에서 찾아야함.
         List<UUID> userIdList = null;
+        Instant time = null;
         if(channel.getType()==ChannelType.PRIVATE){
-            userIdList = userRepository.findAllUserIdByChannelId(uuid);
+            userIdList = readStatusRepository.findAllUserIdByChannelId(uuid);
+            time = readStatusRepository.findLatestTimeByChannelId(uuid);
         }
         ChannelFindDTO channelFindDTO = new ChannelFindDTO(channel, time, userIdList);
         return channelFindDTO;
@@ -90,7 +94,7 @@ public class BasicChannelService implements ChannelService {
         List<ChannelFindDTO> channelFindDTOList = findAllDTO();
         List<ChannelFindDTO> userChannelFindDTOList = channelFindDTOList.stream()
                 .filter(channelFindDTO ->channelFindDTO.getType().equals(ChannelType.PUBLIC)
-                &&
+                ||
                         (channelFindDTO.getType().equals(ChannelType.PRIVATE) &&
                                 channelFindDTO.isUserExist(userId)))
                 .toList();
@@ -135,7 +139,10 @@ public class BasicChannelService implements ChannelService {
     public void deleteChannel(UUID id) {
         //관련된 도메인 삭제
         messageRepository.deleteByChannelId(id);
-        readStatusService.deleteByChannelId(id);
+        Channel  channel = readChannel(id);
+        if(channel.getType()==ChannelType.PRIVATE){
+            readStatusService.deleteByChannelId(id);
+        }
         channelRepository.delete(id);
     }
 }
